requirements.txt
Django==3.2
djangorestframework==3.12.4
flake8==3.6.0
psycopg2==2.7.5